Dry - Clean code principle
