+++ src/app/auth/authSlice.ts
@@ -1,6 +1,7 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit'
 import { onSocialLogin, onLogout, onRegister, onRegisterActivate, onLogin, onRefreshJWT, onCreateProfile, onUpdateProfile, onGetCurrentUserData, onUpdateProfilePicture } from './thunks';
-import { ICurrentUser } from '../../pages/auth/types/interfaces';
+import { ICurrentUser } from '../types/interfaces';
+
 
 export interface AuthState {
   access: string | null;
