+++ src/app/auth/thunks.ts
@@ -1,7 +1,6 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { axios_base } from '../../api/axios_base';
-import { ICurrentUser, IJWTRefreshSuccess, ILoginSuccess, IRegisterSuccess, ISocialLoginSuccess } from '../../pages/auth/types/interfaces';
-import { IOnRegister, ISocialOnLogin, IOnRegisterActivate, IOnLogin, IOnRefreshJWT, IOnCreateProfile, IOnUpdateProfile, IOnGetCurrentUserData, IOnUpdateProfilePicture } from '../types/interfaces';
+import { IOnRegister, ISocialOnLogin, IOnRegisterActivate, IOnLogin, IOnRefreshJWT, IOnCreateProfile, IOnUpdateProfile, IOnGetCurrentUserData, IOnUpdateProfilePicture, ISocialLoginSuccess, ILoginSuccess, IRegisterSuccess, IJWTRefreshSuccess, ICurrentUser } from '../types/interfaces';
 import { AxiosError } from 'axios';
 import { setErrors } from './authSlice';
 import { errorNotification, successNotification } from '../../components/common/Alerts';
@@ -82,7 +81,7 @@ export const onRegisterActivate = createAsyncThunk(
 
 export const onRefreshJWT = createAsyncThunk(
     'auth/onRefreshJWT',
-    async (data: IOnRefreshJWT,  { dispatch, getState }) => {
+    async (data: IOnRefreshJWT,  { dispatch }) => {
         try {
             // check if refresh JWT is valid and refresh token (new access JWT)
             const resp = await axios_base.post<IJWTRefreshSuccess>(`auth/jwt/refresh`, data);
